Move kafka-jwt-auth to devon4j-security-jwt
