+++ modules/security-jwt/pom.xml
@@ -33,18 +33,8 @@
       <artifactId>spring-security-jwt</artifactId>
     </dependency>
     <dependency>
-        <groupId>io.jsonwebtoken</groupId>
-        <artifactId>jjwt</artifactId>
-    </dependency>
-    <!-- Spring Web & Servlet necessary for Request Filters -->
-    <dependency>
-      <groupId>org.springframework</groupId>
-      <artifactId>spring-web</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>javax.servlet</groupId>
-      <artifactId>javax.servlet-api</artifactId>
-      <scope>provided</scope>
+      <groupId>io.jsonwebtoken</groupId>
+      <artifactId>jjwt</artifactId>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
@@ -56,6 +46,31 @@
       <artifactId>devon4j-test</artifactId>
       <scope>test</scope>
     </dependency>
+
+    <!-- Spring Web & Servlet necessary for Request Filters -->
+    <dependency>
+      <groupId>org.springframework</groupId>
+      <artifactId>spring-web</artifactId>
+      <optional>true</optional>
+    </dependency>
+    <dependency>
+      <groupId>javax.servlet</groupId>
+      <artifactId>javax.servlet-api</artifactId>
+      <optional>true</optional>
+    </dependency>
+
+    <!-- Kafka necessary for JwtAuth-Aspect -->
+    <dependency>
+      <groupId>org.springframework.kafka</groupId>
+      <artifactId>spring-kafka</artifactId>
+      <optional>true</optional>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework</groupId>
+      <artifactId>spring-aspects</artifactId>
+      <optional>true</optional>
+    </dependency>
   </dependencies>
+
 
 </project>
