+++ starters/starter-security-jwt/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,1 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.devonfw.module.security.jwt.common.impl.JwtAutoConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.devonfw.module.security.jwt.common.impl.JwtAutoConfiguration,com.devonfw.module.security.jwt.common.impl.JwtAutoConfigurationWeb
